Update timer graphics
Moving to python3 and using new pillow library. The initial impetus for
this was to make the graphics better for a Retina display. However, the
new library also facilitates some better drawing, so it has some side
benefits.

Update the sprites so they don't look horrible, and vertically align the
two widgets' middles.

While I was in there, I updated the 'fast-forward' logic in the
sparklines to use a binary search, so that should speed things up a
little.